requirements.txt
fastapi==0.95.2
motor==3.1.1
python-dotenv==1.0.0
uvicorn==0.22.0
